Add unit tests for mutation rate validation and behavior in Vehicle class
